requirements.txt
click==8.3.1
colorama==0.4.6
dlib-bin==20.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.2.6
opencv-python==4.13.0.92
pillow==12.1.1
